# backend/docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/auth/register:
    post:
      operationId: AuthController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: Account created; OTP dispatched
      summary: Register a citizen account
      tags:
        - Auth
  /api/auth/otp/send:
    post:
      operationId: AuthController_sendOtp
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendOtpDto'
      responses:
        '200':
          description: OTP generated (returned in data for demo)
      summary: Send a 6-digit phone verification OTP
      tags:
        - Auth
  /api/auth/otp/verify:
    post:
      operationId: AuthController_verifyOtp
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpDto'
      responses:
        '200':
          description: ''
      summary: Verify phone OTP
      tags:
        - Auth
  /api/auth/pin/setup:
    post:
      operationId: AuthController_setupPin
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupPinDto'
      responses:
        '200':
          description: ''
      summary: Set or update a 4-digit security PIN
      tags:
        - Auth
  /api/auth/biometric:
    post:
      operationId: AuthController_biometric
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BiometricDto'
      responses:
        '200':
          description: Face ID registered or skipped
      security:
        - access-token: []
      summary: Face ID registered / skip for now
      tags:
        - Auth
  /api/auth/pin/forgot:
    post:
      operationId: AuthController_forgotPin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPinDto'
      responses:
        '200':
          description: OTP dispatched if the phone exists
      summary: Request a PIN-reset OTP by phone
      tags:
        - Auth
  /api/auth/pin/verify-otp:
    post:
      operationId: AuthController_verifyPinResetOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyPinResetDto'
      responses:
        '200':
          description: PIN-reset OTP verified
      summary: Verify PIN-reset OTP
      tags:
        - Auth
  /api/auth/pin/reset:
    post:
      operationId: AuthController_resetPin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPinDto'
      responses:
        '200':
          description: PIN updated
      summary: Set a new 4-digit PIN with a verified OTP
      tags:
        - Auth
  /api/auth/pin/verify:
    post:
      operationId: AuthController_verifyPin
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyPinDto'
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Verify PIN for delete account / logout
      tags:
        - Auth
  /api/auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: ''
        '401':
          description: Invalid credentials
      summary: Login with email/phone + password or PIN
      tags:
        - Auth
  /api/auth/password/forgot:
    post:
      operationId: AuthController_forgotPassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordDto'
      responses:
        '200':
          description: ''
      summary: Request a password-reset OTP
      tags:
        - Auth
  /api/auth/password/verify-otp:
    post:
      operationId: AuthController_verifyResetOtp
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyResetOtpDto'
      responses:
        '200':
          description: ''
      summary: Verify password-reset OTP
      tags:
        - Auth
  /api/auth/password/reset:
    post:
      operationId: AuthController_resetPassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordDto'
      responses:
        '200':
          description: ''
      summary: Set a new password with a verified OTP
      tags:
        - Auth
  /api/auth/me:
    get:
      operationId: AuthController_me
      parameters: []
      responses:
        '200':
          description: ''
        '401':
          description: Missing or invalid Bearer token
      security:
        - access-token: []
      summary: Current user, groups, and active alerts
      tags:
        - Auth
  /api/auth/logout:
    post:
      operationId: AuthController_logout
      responses:
        '200':
          description: Session token revoked
      security:
        - access-token: []
      summary: Log out — invalidate the current session token
      tags:
        - Auth
...
  /api/uploads/images:
    post:
      operationId: UploadsController_upload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
              required:
                - files
      responses:
        '201':
          description: Images stored in S3 (or local disk when AWS_S3_BUCKET is unset)
      summary: Upload profile photos to S3
      tags:
        - Uploads
  /api/uploads/presign:
    post:
      operationId: UploadsController_presign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresignUploadDto'
      responses:
        '200':
          description: Presigned PUT URL
        '503':
          description: S3 is not configured
      summary: Presigned S3 PUT for direct upload
      tags:
        - Uploads
...
  /api/dashboard/users:
    get:
      operationId: DashboardController_users
      parameters:
        - name: q
          required: false
          in: query
          schema:
            example: sarah
            type: string
      responses:
        '200':
          description: ''
      summary: List users (optional search)
      tags:
        - Dashboard
  /api/dashboard/users/{id}/verify:
    patch:
      operationId: DashboardController_toggleVerify
      parameters:
        - name: id
          required: true
          in: path
          schema:
            example: usr-sarah-101
            type: string
      responses:
        '200':
          description: ''
        '404':
          description: User not found
      summary: Toggle user verification
      tags:
        - Dashboard
  /api/dashboard/emergency-types:
    get:
      operationId: DashboardController_emergencyTypes
      parameters: []
      responses:
        '200':
          description: ''
      summary: List emergency types for admin
      tags:
        - Dashboard
    post:
      operationId: DashboardController_createEmergencyType
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmergencyTypeDto'
      responses:
        '201':
          description: ''
      summary: Create an emergency type
      tags:
        - Dashboard
  /api/dashboard/emergency-types/{id}/toggle:
    patch:
      operationId: DashboardController_toggleEmergencyType
      parameters:
        - name: id
          required: true
          in: path
          schema:
            example: et-assault
            type: string
      responses:
        '200':
          description: ''
        '404':
          description: Type not found
      summary: Toggle emergency type active flag
      tags:
        - Dashboard
...
  /api/dashboard/journals:
    get:
      operationId: DashboardController_journals
      responses:
        '200':
          description: Historical journals for all users
      security:
        - access-token: []
      summary: Historical incident journals (all users)
      tags:
        - Dashboard
  /api/emergency-types:
    get:
      operationId: CatalogController_emergencyTypes
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
            example: fire
      responses:
        '200':
          description: Emergency Type grid ordered for the Figma picker
      summary: Emergency Type catalog
      tags:
        - Catalog
...
  /api/journals:
    get:
      operationId: JournalsController_list
      responses:
        '200':
          description: Incident Journal list for the authenticated user
      security:
        - access-token: []
      summary: Incident Journal list
      tags:
        - Journal
    post:
      operationId: JournalsController_create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJournalDto'
      responses:
        '201':
          description: Saved journal entry
      security:
        - access-token: []
      summary: Save a New Entry
      tags:
        - Journal
...
info:
  title: SafeAlert Emergency API
  description: >-
    REST + Socket.io API for the SafeAlert emergency response system.


    **Demo accounts**

    - Super Admin: `admin@safealert.app` / `adminpassword`

    - Ops: `ops@safealert.app` / `opspassword`

    - Citizen (Sarah): `sarah.johnson@example.com` / PIN `1234`

    - Phone OTP for Sarah: `123456`


        Use **Authorize** after login and paste the JWT as a Bearer token for
        `/api/auth/me`.

        **Socket.io** (`ws://localhost:5002`, same origin as REST)

        - Client → server: `alert:join`, `alert:telemetry`, `alert:message:send`,
        `alert:quick_response`, `admin:subscribe`

        - Server → client: `alert:state`, `alert:telemetry:update`,
        `alert:messages:update`, `alert:resolved`, `notification:new`,
        `admin:alert:new`, `admin:alert:resolved`, `admin:alert:telemetry`,
        `admin:metrics`
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: JWT from POST /api/auth/login
  schemas:
    RegisterDto:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
          example: Jordan Lee
        email:
          type: string
          format: email
          example: jordan.lee@example.com
        phone:
          type: string
          minLength: 7
          example: +1 (555) 321-0987
        dob:
          type: string
          example: '1998-05-14'
        race:
          type: string
          example: White
        location:
          type: string
          example: New York, NY
        emergencyContactName:
          type: string
          minLength: 2
          example: James Johnson
        emergencyContactPhone:
          type: string
          minLength: 7
          example: +1 (555) 987-6543
        emergencyContactRelation:
          type: string
          minLength: 2
          example: Father
        profilePhotos:
          example:
            - https://images.unsplash.com/photo-1494790108377-be9c29b29330?w=300
          type: array
          items:
            type: string
        pin:
          type: string
          pattern: ^\d{4}$
          example: '1234'
          description: 4-digit de-escalation PIN
        password:
          type: string
          minLength: 6
          example: password123
      required:
        - fullName
        - email
        - phone
        - emergencyContactName
        - emergencyContactPhone
        - emergencyContactRelation
    SendOtpDto:
      type: object
      properties:
        phone:
          type: string
          minLength: 7
          example: +1 (555) 234-5678
      required:
        - phone
    VerifyOtpDto:
      type: object
      properties:
        phone:
          type: string
          minLength: 7
          example: +1 (555) 234-5678
        code:
          type: string
          minLength: 6
          maxLength: 6
          example: '123456'
          description: Demo code for Sarah's phone is 123456
      required:
        - phone
        - code
    SetupPinDto:
      type: object
      properties:
        pin:
          type: string
          pattern: ^\d{1,4}$
          example: '1234'
      required:
        - pin
    BiometricDto:
      type: object
      properties:
        enabled:
          type: boolean
          example: true
      required:
        - enabled
    ForgotPinDto:
      type: object
      properties:
        phone:
          type: string
          minLength: 7
          example: +1 (555) 234-5678
      required:
        - phone
    VerifyPinResetDto:
      type: object
      properties:
        phone:
          type: string
          minLength: 7
          example: +1 (555) 234-5678
        code:
          type: string
          minLength: 6
          maxLength: 6
          example: '123456'
      required:
        - phone
        - code
    ResetPinDto:
      type: object
      properties:
        phone:
          type: string
          minLength: 7
          example: +1 (555) 234-5678
        code:
          type: string
          minLength: 6
          maxLength: 6
          example: '123456'
        newPin:
          type: string
          pattern: ^\d{1,4}$
          example: '1234'
      required:
        - phone
        - code
        - newPin
    VerifyPinDto:
      type: object
      properties:
        userId:
          type: string
          example: usr-sarah-101
          description: Admin-only override. Citizens always verify their own PIN.
        pin:
          type: string
          pattern: ^\d{1,4}$
          example: '1234'
      required:
        - pin
    LoginDto:
      type: object
      properties:
        emailOrPhone:
          type: string
          minLength: 3
          example: admin@safealert.app
          description: Email address or phone number
        pin:
          type: string
          pattern: ^\d{4}$
          example: '1234'
          description: Citizen 4-digit PIN. Send this or password, not both required.
        password:
          type: string
          minLength: 4
          example: adminpassword
          description: >-
            Admin / password login. Use adminpassword or opspassword for demo
            admins.
      required:
        - emailOrPhone
    ForgotPasswordDto:
      type: object
      properties:
        email:
          type: string
          format: email
          example: admin@safealert.app
      required:
        - email
    VerifyResetOtpDto:
      type: object
      properties:
        email:
          type: string
          format: email
          example: admin@safealert.app
        code:
          type: string
          minLength: 6
          maxLength: 6
          pattern: ^\d{6}$
          example: '123456'
      required:
        - email
        - code
    ResetPasswordDto:
      type: object
      properties:
        email:
          type: string
          format: email
          example: admin@safealert.app
        code:
          type: string
          minLength: 6
          maxLength: 6
          pattern: ^\d{6}$
          example: '123456'
        newPassword:
          type: string
          minLength: 6
          example: adminpassword
      required:
        - email
        - code
        - newPassword
...
    CreateEmergencyTypeDto:
      type: object
      properties:
        key:
          type: string
          example: FLOOD
        label:
          type: string
          example: Flood / Water Rescue
        severity:
          enum:
            - LOW
            - URGENT
            - HIGH
            - CRITICAL
          type: string
          example: HIGH
        icon:
          type: string
          example: Waves
        description:
          type: string
          example: Rapid water rise or trapped occupants
        isActive:
          type: boolean
          example: true
      required:
        - key
        - label
        - severity
        - icon
        - description
    CreateJournalDto:
      type: object
      properties:
        body:
          type: string
          example: Was followed home from the subway. Got home safely
        content:
          type: string
          description: Alias of body (Figma composer)
        type:
          type: string
          example: Incident
          description: Incident | Test | Update (defaults to Update)
...